Removed `MouseHandler` Class
Moved its functionality entirely into `TranscriptionViewController`.
